requirement.txt
flask==1.1.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==0.15.5
keras==2.3.1
tensorflow-gpu==2.1.0
numpy>=1.18
scipy>=0.15.1
